+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 * Added player roles message when the game is force ended. ([#225](https://github.com/hjwylde/werewolf/issues/225))
 * Added the no-role-reveal variant. ([#181](https://github.com/hjwylde/werewolf/issues/181))
 * Added a public start-of-game message for the Spiteful Ghost. ([#220](https://github.com/hjwylde/werewolf/issues/220))
+* Removed the public start-of-game message for the Fallen Angel. ([#229](https://github.com/hjwylde/werewolf/issues/229))
 
 ### v1.2.0.2